........
  r11522 | p_lindheimer | 2011-02-21 07:52:37 -0800 (Mon, 21 Feb 2011) | 1 line
  
  don't include robots.txt in framework upgrade as ownership is not guaranteed re #4844
........

+++ publish.pl
@@ -97,6 +97,9 @@ while ($moddir = shift @ARGV) {
 		if (system("rm -rf $framework/htdocs/index.html")) {
 			die "FATAL: failed to trim htdocs/index.html\n";
 		}
+		if (system("rm -rf $framework/htdocs/robots.txt")) {
+			die "FATAL: failed to trim htdocs/robots.txt\n";
+		}
 		if (system("rm -rf $framework/htdocs/admin/modules/_cache")) {
 			die "FATAL: failed to trim modules/_cache\n";
 		}
